Added: Score Pickup, pickup de-spawn, score UI. Note (Pickup de-spawn on collision is commented out until I find a better solution)
Changes Description:
Score Pickup - On collision between the player and enemy, the enemy will de-spawn and make an pickup object. (To be changed once combat is included)

Pickup Despawn - If the pickup is left for x amount of time, it will delete itself.

Score UI - Score changes based on points received